Oops, this needs to be consistent
git-svn-id: https://llvm.org/svn/llvm-project/poolalloc/trunk@17418 91177308-0d34-0410-b5e6-96231b3b80d8

+++ runtime/FL2Allocator/PoolAllocator.h
@@ -47,7 +47,7 @@ struct LargeArrayHeader {
   
   // Marker: this is the ObjectSize marker which MUST BE THE LAST ELEMENT of
   // this header!
-  unsigned Marker;
+  unsigned long Marker;
 };
